requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-google